move cardinality, percentiles, and statistical into legacy
