+++ src/example-types/index.js
@@ -16,12 +16,11 @@ module.exports = {
 
   // Single Metrics
   results: require('./results'),
-  cardinality: require('./cardinality'),
-  percentiles: require('./percentiles'),
-  statistical: require('./statistical'),
+  cardinality: require('./legacy/cardinality'),
+  percentiles: require('./legacy/percentiles'),
   stats: require('./metricGroups/stats'),
   valuesDelta: require('./metricGroups/valuesDelta'),
-
+  
   // Metric Groups
   groupedByFieldValue: require('./metricGroups/groupedByFieldValue'),
   groupedByPercentiles: require('./metricGroups/groupedByPercentiles'),
@@ -29,8 +28,9 @@ module.exports = {
   groupedByNumberRanges: require('./metricGroups/groupedByNumberRanges'),
   groupedByNumberInterval: require('./metricGroups/groupedByNumberInterval'),
   groupedByValuePartition: require('./metricGroups/groupedByValuePartition'),
-
+  
   // Legacy (covered by metric groups)
+  statistical: require('./legacy/statistical'),
   terms_stats: require('./legacy/terms_stats'),
   dateHistogram: require('./legacy/dateHistogram'),
   rangeStats: require('./legacy/rangeStats'),
