+++ src/example-types/legacy/statistical.js
