+++ test/example-types/legacy/cardinality.js
@@ -1,4 +1,4 @@
-let sequentialResultTest = require('./testUtils').sequentialResultTest
+let sequentialResultTest = require('../testUtils').sequentialResultTest
 
 describe('cardinality', () => {
   let statsTest = (...x) =>
