+++ test/example-types/legacy/percentiles.js
@@ -1,4 +1,4 @@
-let sequentialResultTest = require('./testUtils').sequentialResultTest
+let sequentialResultTest = require('../testUtils').sequentialResultTest
 
 describe('percentiles', () => {
   let percentilesTest1 = (...x) =>
